_root/_ui/src/requirements.txt
streamlit
pandas
requests
python-dotenv 
##*****Install this manually*****
#st-aggrid